#368: Moved the SettingProperties model to its own `settings` app
