+++ oppia/settings/admin.py
@@ -1,0 +1,4 @@
+from oppia.settings.models import SettingProperties
+from django.contrib import admin
+
+admin.site.register(SettingProperties)
